+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [#3] Improve status conditions for ServiceAccountRequests
+- [#3] Simplify serviceaccount-request params to just an array of strings
 
 ## [v1.0.0] - 2026-04-22
 ### Added